moved SupportTextClassification from interfaces.py into score directory to avoid circular import error
