+++ pyrit/score/__init__.py
@@ -1,6 +1,8 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
+from pyrit.score.score_class import Score
+from pyrit.score.support_text_classification import SupportTextClassification
 from pyrit.score.engine import evaluate, evaluate_async, score_text
 from pyrit.score.gandalf_classifier import GandalfScorer, GandalfBinaryScorer
 from pyrit.score.gpt_classifier import (
@@ -26,9 +28,10 @@ from pyrit.score.gpt_classifier import (
 from pyrit.score.markdown_injection import MarkdownInjectionClassifier
 from pyrit.score.text import TextConversationTokenFinder
 from pyrit.score.substring_scorer import SubStringScorer
-from pyrit.score.score import Score
 
 __all__ = [
+    "Score",
+    "SupportTextClassification",
     "score_text",
     "evaluate",
     "evaluate_async",
@@ -55,5 +58,4 @@ __all__ = [
     "MarkdownInjectionClassifier",
     "TextConversationTokenFinder",
     "SubStringScorer",
-    "Score"
 ]
