+++ pyrit/score/gandalf_classifier.py
@@ -8,8 +8,7 @@ from openai import BadRequestError
 import uuid
 
 from pyrit.completion.gandalf_completion import GandalfLevel
-from pyrit.interfaces import SupportTextClassification
-from pyrit.score import Score
+from pyrit.score import Score, SupportTextClassification
 from pyrit.models import PromptRequestPiece, PromptRequestResponse
 from pyrit.prompt_target import PromptChatTarget
