+++ pyrit/score/gpt_classifier.py
@@ -9,8 +9,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Union
 
-from pyrit.interfaces import SupportTextClassification
-from pyrit.score import Score
+from pyrit.score import Score, SupportTextClassification
 from pyrit.models import PromptRequestPiece, PromptRequestResponse, PromptTemplate
 from pyrit.prompt_target import PromptChatTarget
 from pyrit.common.path import PYRIT_PATH
