+++ pyrit/score/support_text_classification.py
@@ -1,0 +1,14 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+
+import abc
+from abc import abstractmethod
+
+from pyrit.score import Score
+
+
+class SupportTextClassification(abc.ABC):
+    @abstractmethod
+    def score_text(self, text: str) -> Score:
+        """Score the text and return a Score object."""
+        raise NotImplementedError("score_text method not implemented")
